working on 32k clicking int to long

+++ src/main/java/com/osroyale/ItemPile.java
@@ -8,6 +8,6 @@ final class ItemPile {
 	Renderable aClass30_Sub2_Sub4_48;
 	Renderable aClass30_Sub2_Sub4_49;
 	Renderable aClass30_Sub2_Sub4_50;
-	int uid;
+	long uid;
 	int anInt52;
 }
